requirements.txt
requests
python-dotenv
cachetools
ratelimit
Jinja2
